small refactor of choice selection
