+++ src/choice/choice_style.rs
@@ -1,0 +1,6 @@
+use crate::templates::{Choice, TemplateVariable};
+use crate::error::ZatResult;
+
+pub trait ChoiceStyle {
+  fn get_choice<'a>(variable: &TemplateVariable, items: &'a [&'a Choice]) -> ZatResult<&'a Choice>;
+}
