+++ src/choice/mod.rs
@@ -5,9 +5,17 @@ pub mod choice_error;
 pub mod choice_scope_filter;
 pub mod default_choice_scope_filter;
 
+mod numbered_choice_style;
+mod selection_choice_style;
+mod choice_style;
+
 pub use choice_runner::ChoiceRunner;
 pub use selected_choices::SelectedChoices;
 pub use default_choice_runner::DefaultChoiceRunner;
 pub use choice_error::ChoiceError;
 pub use choice_scope_filter::ChoiceScopeFilter;
 pub use default_choice_scope_filter::DefaultChoiceScopeFilter;
+
+use numbered_choice_style::NumberedChoiceStyle;
+use selection_choice_style::SelectionChoiceStyle;
+use choice_style::ChoiceStyle;
